remove ES2018 transforms for ES2018 target
