+++ src/compiler/tsconfig.json
@@ -30,6 +30,7 @@
         "transformers/destructuring.ts",
         "transformers/ts.ts",
         "transformers/es2017.ts",
+        "transformers/es2018.ts",
         "transformers/esnext.ts",
         "transformers/jsx.ts",
         "transformers/es2016.ts",
